requirements.txt
symphony-bdk-python>=2.5.0
Jinja2~=3.0
